requirements.txt
python-docx
openpyxl
mammoth
PyMuPDF
python-pptx
patool
pytesseract
pillow
pymupdf
pandas
pdf2image
opencv-python
regex